Dev: network_utils: Refactor bootstrap network helpers
Move bootstrap-specific network helpers into crmsh.network_utils:

- Move bootstrap network input validation into BootstrapValidation
- Move remote cluster-node IP discovery into network_utils
- Move firewalld service management into network_utils
- Keep pure SSH port and reachable-node filtering in network_utils
- Keep cluster reachability orchestration in utils.py where xmlutil is already used
- Keep bootstrap aliases for existing call sites except get_cluster_node_ips
- Update unit-test patches for the new module paths
- Anchor report unit-test fixture lookup to the test directory
- Remove stale network_utils imports from UI and feature-step files
