+++ test/unittests/test_network_utils.py
@@ -9,7 +9,7 @@ from unittest import mock
 
 import pytest
 
-from crmsh import network_utils
+from crmsh import network_utils, utils
 
 
 @mock.patch("crmsh.network_utils.sh.LocalShell.get_rc_and_error")
@@ -302,7 +302,7 @@ class TestInterfacesInfo(unittest.TestCase):
 
 
 @mock.patch('crmsh.network_utils.ssh_port_reachable_check')
-@mock.patch("crmsh.network_utils.xmlutil.CrmMonXmlParser")
+@mock.patch("crmsh.utils.xmlutil.CrmMonXmlParser")
 def test_check_all_nodes_reachable_dead_nodes(mock_xml, mock_reachable):
     mock_xml_inst = mock.Mock()
     mock_xml.return_value = mock_xml_inst
@@ -310,15 +310,15 @@ def test_check_all_nodes_reachable_dead_nodes(mock_xml, mock_reachable):
     mock_xml_inst.get_node_list.side_effect = [["node1"], ["node2"]]
     mock_reachable.side_effect = ValueError
 
-    with pytest.raises(network_utils.DeadNodeError) as err:
-        network_utils.check_all_nodes_reachable("testing")
+    with pytest.raises(utils.DeadNodeError) as err:
+        utils.check_all_nodes_reachable("testing")
     assert err.value.summary.dead_nodes == ["node2"]
 
 
 @mock.patch('crmsh.network_utils.check_ssh_passwd_need')
-@mock.patch("crmsh.network_utils.crmsh.user_of_host.UserOfHost.instance")
+@mock.patch("crmsh.utils.crmsh.user_of_host.UserOfHost.instance")
 @mock.patch('crmsh.network_utils.ssh_port_reachable_check')
-@mock.patch("crmsh.network_utils.xmlutil.CrmMonXmlParser")
+@mock.patch("crmsh.utils.xmlutil.CrmMonXmlParser")
 def test_check_all_nodes_reachable(mock_xml, mock_reachable, mock_user_of_host, mock_check_passwd):
     mock_xml_inst = mock.Mock()
     mock_xml.return_value = mock_xml_inst
@@ -328,7 +328,7 @@ def test_check_all_nodes_reachable(mock_xml, mock_reachable, mock_user_of_host, 
     mock_user_of_host.return_value = mock_user_of_host_inst
     mock_user_of_host_inst.user_pair_for_ssh = mock.Mock(return_value=("root", "root"))
     mock_check_passwd.return_value = False
-    network_utils.check_all_nodes_reachable("testing")
+    utils.check_all_nodes_reachable("testing")
     mock_reachable.assert_called_once_with("node1")
 
 
@@ -343,5 +343,4 @@ def test_get_reachable_node_list(mock_reachable, mock_warn):
         mock.call("node2"),
         mock.call("node3")
     ])
-
